Refactor RevocationInfo into a separate module
Required for downstream revocation propagation.

https://github.com/netsec-ethz/scion/issues/108
